Move network tools from genesis module to the Norma module
